move bookmark menu to subfeature of profile feature
