--- lib/features/bookmark/bindings/bookmark_binding.dart
@@ -1,9 +1,0 @@
-import 'package:dolang/features/bookmark/controllers/bookmark_controller.dart';
-import 'package:get/get.dart';
-
-class BookmarkBinding extends Bindings {
-  @override
-  void dependencies() {
-    Get.put(BookmarkController());
-  }
-}
