--- lib/features/bookmark/constants/bookmark_api_constant.dart
@@ -1,5 +1,0 @@
-class BookmarkApiConstant {
-  final String destination = 'destination';
-  final String user = 'users';
-  final String bookmark = 'bookmark';
-}
