--- lib/features/bookmark/constants/bookmark_assets_constant.dart
@@ -1,3 +1,0 @@
-    class BookmarkAssetsConstant {
-    final String example = 'example';
-    }
