+++ lib/features/dashboard/bindings/dashboard_binding.dart
@@ -1,7 +1,7 @@
-import 'package:dolang/features/bookmark/controllers/bookmark_controller.dart';
 import 'package:dolang/features/dashboard/controllers/dashboard_controller.dart';
 import 'package:dolang/features/home/controllers/home_controller.dart';
 import 'package:dolang/features/profile/controllers/profile_controller.dart';
+import 'package:dolang/features/profile/sub_features/bookmark/controllers/profile_bookmark_controller.dart';
 import 'package:get/get.dart';
 
 class DashboardBinding extends Bindings {
