+++ lib/features/profile/constants/profile_api_constant.dart
@@ -1,3 +1,5 @@
-    class ProfileApiConstant {
-    final String example = 'example';
-    }
+class ProfileApiConstant {
+  final String destination = 'destination';
+  final String user = 'users';
+  final String bookmark = 'bookmark';
+}
