+++ lib/features/profile/view/ui/profile_screen.dart
@@ -1,6 +1,9 @@
 import 'package:dolang/features/profile/constants/profile_assets_constant.dart';
+import 'package:dolang/features/profile/sub_features/bookmark/view/components/bookmark_conditional_list.dart';
 import 'package:dolang/features/profile/view/components/log_out_button_component.dart';
+import 'package:dolang/features/profile/view/components/profile_icon_component.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class ProfileScreen extends StatelessWidget {
   ProfileScreen({super.key});
@@ -8,9 +11,36 @@ class ProfileScreen extends StatelessWidget {
   final assetsConstant = ProfileAssetsConstant();
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: LogOutButtonComponent(),
+    return Scaffold(
+      body: CustomScrollView(
+        slivers: [
+          const SliverAppBar(
+            expandedHeight: kToolbarHeight,
+            floating: false,
+            pinned: true,
+            flexibleSpace: FlexibleSpaceBar(
+              title: Text('Profile'),
+              // background: Image(
+              //   image: AssetImage('assets/images/profile.jpg'),
+              //   fit: BoxFit.cover,
+              // ),
+            ),
+          ),
+          SliverList(
+            delegate: SliverChildListDelegate(
+              [
+                const ProfileIconComponent(),
+                const BookmarkConditionalList(),
+                Padding(
+                  padding: EdgeInsets.symmetric(
+                    horizontal: 16.h,
+                  ),
+                  child: const LogOutButtonComponent(),
+                ),
+              ],
+            ),
+          ),
+        ],
       ),
     );
   }
